src/webapp/requirements.txt
fastapi
pillow
python-multipart
torch
torchvision
uvicorn
